+++ src/components/ApplicationPage/Application/index.tsx
@@ -2,21 +2,13 @@ import { MouseEvent, useState } from 'react';
 import ApplyCalendar from '../ApplyCalendar';
 import * as S from './styles';
 import { DateSelectArg } from '@fullcalendar/core/index.js';
+import { DateClickArg } from '@fullcalendar/interaction/index.js';
+import ApplyInfo from '../ApplyInfo';
 
 function Application() {
   const [select, setSelect] = useState<'annual' | 'duty'>('annual');
   const [date, setDate] = useState({ start_date: '', end_date: '' });
 
-  const selectText = select === 'annual' ? '연차' : '당직';
-  const selectDays = (() => {
-    if (!date.start_date || !date.end_date) return 0;
-    const startDate = new Date(date.start_date);
-    const endDate = new Date(date.end_date);
-    const oneDay = 1000 * 60 * 60 * 24; // milliseconds in a day
-    const diffTime = Math.abs(endDate.getTime() - startDate.getTime());
-    const diffDays = Math.ceil(diffTime / oneDay);
-    return diffDays + 1; // include start date
-  })();
   const handleClick = (e: MouseEvent<HTMLDivElement>) => {
     const target = e.target as HTMLButtonElement;
     const select = target.dataset.select as 'annual' | 'duty';
@@ -24,10 +16,13 @@ function Application() {
       setSelect(select);
     }
   };
-  const handleDateSelect = (date: DateSelectArg) => {
-    if (select === 'duty') return;
-    const { startStr, endStr } = date;
-    console.log(startStr, endStr);
+  const handleDateSelect = (date: DateSelectArg | DateClickArg) => {
+    if (select === 'duty') {
+      const { dateStr } = date as DateClickArg;
+      setDate({ start_date: dateStr, end_date: dateStr });
+      return;
+    }
+    const { startStr, endStr } = date as DateSelectArg;
     setDate({
       start_date: startStr,
       // fullCalendar에서 endStr은 선택한 날짜의 다음날을 가리킴 -> 해당 오류 수정
@@ -46,13 +41,7 @@ function Application() {
         </S.SelectBtn>
       </S.SelectBtns>
       <ApplyCalendar select={select} handleDateSelect={handleDateSelect} />
-      <S.ApplyInfo>
-        <S.ApplyInfoTitle>{selectText} 신청 정보</S.ApplyInfoTitle>
-        <S.ApplyInfoContent>
-          신청 날짜 : {date.start_date} ~ {date.end_date}
-        </S.ApplyInfoContent>
-        <S.ApplyInfoContent>신청 일수 : {selectDays}일</S.ApplyInfoContent>
-      </S.ApplyInfo>
+      {date.start_date && date.end_date && <ApplyInfo select={select} date={date} />}
     </S.Wrapper>
   );
 }
